+++ src/test/java/jpa/repository/LineRepositoryTest.java
@@ -1,16 +1,12 @@
 package jpa.repository;
 
 import jpa.domain.Line;
-import jpa.domain.Station;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -20,21 +16,6 @@ class LineRepositoryTest {
 
 	@Autowired
 	LineRepository lines;
-
-	@Autowired
-	StationRepository stationRepository;
-
-	@BeforeEach
-	void setUp() {
-		Line line = new Line("신분당선", "red");
-		lines.save(line);
-		Set<Station> stations = new HashSet<>();
-		stations.add(new Station("강남역"));
-		stations.add(new Station("판교역"));
-
-		line.addAllStation(stations);
-		this.stationRepository.saveAll(stations);
-	}
 
 	@DisplayName("라인 저장 테스트")
 	@Test
@@ -66,12 +47,5 @@ class LineRepositoryTest {
 		List<Line> red = lines.findByColor("red");
 
 		assertThat(red.size()).isEqualTo(2);
-	}
-
-	@DisplayName("노선 조회 시 속한 지하철역을 볼 수 있는 테스트")
-	@Test
-	void 노선조회_TEST() {
-		Line 신분당선 = lines.findByName("신분당선");
-		assertThat(신분당선.getStation().size()).isEqualTo(2);
 	}
 }
